server/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
asyncpg==0.29.0
pydantic==2.11.7
python-multipart==0.0.20
pysnmp==4.4.12
python-dotenv==1.0.0
alembic==1.13.1
aiohttp==3.9.1 